@@ -1,59 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Runtime.InteropServices;
+using System.Linq;
 using Dalamud.Game.Text.SeStringHandling;
 using Dalamud.Game.Text.SeStringHandling.Payloads;
+using Dalamud.Memory;
 
 namespace PartyIcons.Utils;
 
 public static class SeStringUtils
 {
-    public static IntPtr emptyPtr;
-
-    public static void Initialize()
-    {
-        emptyPtr = SeStringToPtr(Text(""));
-    }
-
-    public static void Dispose() { }
-
-    public static SeString SeStringFromPtr(IntPtr seStringPtr)
-    {
-        byte b;
-        var offset = 0;
-
-        unsafe
-        {
-            while ((b = *(byte*) (seStringPtr + offset)) != 0)
-            {
-                offset++;
-            }
-        }
-
-        var bytes = new byte[offset];
-        Marshal.Copy(seStringPtr, bytes, 0, offset);
-
-        return SeString.Parse(bytes);
-    }
-
-    public static IntPtr SeStringToPtr(SeString seString)
-    {
-        var bytes = seString.Encode();
-        var pointer = Marshal.AllocHGlobal(bytes.Length + 1);
-        Marshal.Copy(bytes, 0, pointer, bytes.Length);
-        Marshal.WriteByte(pointer, bytes.Length, 0);
-
-        return pointer;
-    }
-
-    public static void FreePtr(IntPtr seStringPtr)
-    {
-        if (seStringPtr != emptyPtr)
-        {
-            Marshal.FreeHGlobal(seStringPtr);
-        }
-    }
-
     public static SeString Text(string rawText)
     {
         var seString = new SeString(new List<Payload>());
@@ -62,27 +17,22 @@ public static class SeStringUtils
         return seString;
     }
 
-    public static SeString Text(string text, ushort color)
+    // Defaults to glowColor 51 (black glow) for maximum contrast
+    public static SeString Text(string text, ushort fgColor, ushort glowColor = 51)
     {
         var seString = new SeString(new List<Payload>());
-        seString.Append(new UIForegroundPayload(color));
+        seString.Append(new UIGlowPayload(glowColor));
+        seString.Append(new UIForegroundPayload(fgColor));
         seString.Append(new TextPayload(text));
         seString.Append(UIForegroundPayload.UIForegroundOff);
+        seString.Append(UIGlowPayload.UIGlowOff);
 
         return seString;
     }
 
-    public static SeString Icon(BitmapFontIcon icon, string prefix = null)
+    public static string PrintRawStringArg(IntPtr arg)
     {
-        var seString = new SeString(new List<Payload>());
-
-        if (prefix != null)
-        {
-            seString.Append(new TextPayload(prefix));
-        }
-
-        seString.Append(new IconPayload(icon));
-
-        return seString;
+        var seString = MemoryHelper.ReadSeStringNullTerminated(arg);
+        return string.Join("", seString.Payloads.Select(payload => $"[{payload}]"));
     }
 }
